incoporated should migrate function to the migrator
