+++ cmd/migrate/migrate.go
@@ -4,42 +4,13 @@ import (
 	"fmt"
 	"github.com/dymensionxyz/roller/cmd/utils"
 	"github.com/dymensionxyz/roller/config"
-	"github.com/dymensionxyz/roller/sequencer"
 	"github.com/dymensionxyz/roller/version"
-	"github.com/pelletier/go-toml"
 	"github.com/spf13/cobra"
 	"strings"
 )
 
-type VersionMigrator interface {
-	PerformMigration(rlpCfg config.RollappConfig) error
-}
-
-func NewVersionMigrator(version string) VersionMigrator {
-	switch version {
-	case "v0.1.4":
-		return &VersionMigratorV014{}
-	default:
-		return nil
-	}
-}
-
-type VersionData struct {
-	Major int
-	Minor int
-	Patch int
-}
-
-type VersionMigratorV014 struct{}
-
-func (v *VersionMigratorV014) PerformMigration(rlpCfg config.RollappConfig) error {
-	dymintTomlPath := sequencer.GetDymintFilePath(rlpCfg.Home)
-	dymintCfg, err := toml.LoadFile(dymintTomlPath)
-	if err != nil {
-		return err
-	}
-	sequencer.EnableDymintMetrics(dymintCfg)
-	return config.WriteTomlToFile(dymintTomlPath, dymintCfg)
+var migrationsRegistry = []VersionMigrator{
+	&VersionMigratorV014{},
 }
 
 func Cmd() *cobra.Command {
@@ -49,24 +20,12 @@ func Cmd() *cobra.Command {
 		Run: func(cmd *cobra.Command, args []string) {
 			home := cmd.Flag(utils.FlagNames.Home).Value.String()
 			rlpCfg, err := config.LoadConfigFromTOML(home)
-			if err != nil {
-				utils.PrettifyErrorIfExists(err)
-				return
-			}
+			utils.PrettifyErrorIfExists(err)
 			prevVersionData, err := GetVersionData(rlpCfg)
-			if err != nil {
-				utils.PrettifyErrorIfExists(err)
-				return
-			}
-			if prevVersionData.Major < 1 {
-				if prevVersionData.Minor < 2 {
-					if prevVersionData.Patch < 4 {
-						migrator := NewVersionMigrator("v0.1.4")
-						if migrator != nil {
-							err := migrator.PerformMigration(rlpCfg)
-							utils.PrettifyErrorIfExists(err)
-						}
-					}
+			utils.PrettifyErrorIfExists(err)
+			for _, migrator := range migrationsRegistry {
+				if migrator.ShouldMigrate(*prevVersionData) {
+					utils.PrettifyErrorIfExists(migrator.PerformMigration(rlpCfg))
 				}
 			}
 			trimmedCurrentVersion := strings.Split(version.BuildVersion, "-")[0]
@@ -97,4 +56,15 @@ func GetVersionData(rlpCfg config.RollappConfig) (*VersionData, error) {
 		Minor: minor,
 		Patch: patch,
 	}, nil
+}
+
+type VersionMigrator interface {
+	PerformMigration(rlpCfg config.RollappConfig) error
+	ShouldMigrate(prevVersion VersionData) bool
+}
+
+type VersionData struct {
+	Major int
+	Minor int
+	Patch int
 }
